Changed to use buffer instead of unmounting when in background
It seemed that electron/chrome only gives us a fixed time window to exec our logic before app going to background, when app is in background, all API calls to HTML DOM will be queued up and exec them all at once the next time app going to foreground again. This is the main cause of our hanging problem. To summarise, if our unmounting logic didn’t finish executing in that time window when going to background, the app will freeze. This commit makes sure mobx reactions are not triggered at all when in background